pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "cryptography>=46.0.3",
    "dj-database-url>=3.1.0",
    "django>=5.2.10",
    "django-celery-beat>=2.8.1",
    "gunicorn>=23.0.0",
    "o365>=2.1.8",
    "pdfkit>=1.0.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pylibdmtx>=0.1.10",
    "pymupdf>=1.26.7",
    "python-magic>=0.4.27",
    "redis>=7.1.0",
    "whitenoise>=6.11.0",
]
